Modifiquei o index.html

+++ src/main/java/com/example/model/Credito.java
@@ -1,19 +1,19 @@
 package com.example.model;
 
 public class Credito {
-    private int idCred;
+    private Long idCred;
     private String nomCred;
     private double valCred;
     private boolean dizCred;
     
-    public Credito(int idCred, String nomCred, double valCred, boolean dizCred) {
+    public Credito(Long idCred, String nomCred, double valCred, boolean dizCred) {
         this.idCred = idCred;
         this.nomCred = nomCred;
         this.valCred = valCred;
         this.dizCred = dizCred;
     }
 
-    public int getIdCred() { return idCred; }
+    public Long getIdCred() { return idCred; }
     public String getNomCred() { return nomCred; }
     public void setNomCred(String nomCred) { this.nomCred = nomCred; }
     public double getValCred() { return valCred; }
